requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
MouseInfo==0.1.3
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.3
Pillow==8.2.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.7
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
